Add automated validation of subscriptions.json
Building the new validation csproj runs the tests by default by using
the xunit.runner.msbuild NuGet package.
