+++ Maestro/Microsoft.DotNet.Maestro.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Maestro.WebApi", "src\Microsoft.DotNet.Maestro.WebApi\Microsoft.DotNet.Maestro.WebApi.csproj", "{BFEC6AD3-4215-45DF-8313-27218613DA34}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Maestro.WebJob", "src\Microsoft.DotNet.Maestro.WebJob\Microsoft.DotNet.Maestro.WebJob.csproj", "{F685F417-BB64-4346-9FF1-43599DDEFE22}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Maestro", "src\Microsoft.DotNet.Maestro\Microsoft.DotNet.Maestro.csproj", "{DA51EB16-4B3E-4909-89EA-80BD8D017E56}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Maestro.SubscriptionValidation", "test\Microsoft.DotNet.Maestro.SubscriptionValidation\Microsoft.DotNet.Maestro.SubscriptionValidation.csproj", "{8BC073BE-0C4D-47F7-99F8-DF9906C3340C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{DA51EB16-4B3E-4909-89EA-80BD8D017E56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA51EB16-4B3E-4909-89EA-80BD8D017E56}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DA51EB16-4B3E-4909-89EA-80BD8D017E56}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BC073BE-0C4D-47F7-99F8-DF9906C3340C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BC073BE-0C4D-47F7-99F8-DF9906C3340C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BC073BE-0C4D-47F7-99F8-DF9906C3340C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BC073BE-0C4D-47F7-99F8-DF9906C3340C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
