+++ Maestro/src/Microsoft.DotNet.Maestro.WebApi/Microsoft.DotNet.Maestro.WebApi.csproj
@@ -99,13 +99,16 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
+    <Compile Include="Models\ActionList.cs" />
     <Compile Include="Models\Commit.cs" />
     <Compile Include="Models\ModifiedFileModel.cs" />
     <Compile Include="Models\PushWebHookEvent.cs" />
     <Compile Include="Models\Repository.cs" />
+    <Compile Include="Models\Subscription.cs" />
+    <Compile Include="Models\SubscriptionsModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\AzureStorageService.cs" />
-    <Compile Include="Services\EventRegistrationTable.cs" />
+    <Compile Include="Services\HandlerResolver.cs" />
     <Compile Include="Services\VerifyPayloadSignatureAttribute.cs" />
   </ItemGroup>
   <ItemGroup>
