+++ Maestro/src/Microsoft.DotNet.Maestro.WebApi/Models/ActionList.cs
@@ -1,0 +1,32 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace Microsoft.DotNet.Maestro.WebApi.Models
+{
+    public class ActionsList
+    {
+        [JsonExtensionData]
+        private IDictionary<string, JToken> ActionNames { get; set; }
+
+        public JObject GetAction(string name)
+        {
+            JObject action = null;
+
+            if (!string.IsNullOrEmpty(name))
+            {
+                JToken token;
+                if (ActionNames.TryGetValue(name, out token))
+                {
+                    action = token as JObject;
+                }
+            }
+
+            return action;
+        }
+    }
+}
